move lsp functions to their own module
