+++ sway-lsp/src/tests/mod.rs
@@ -1,23 +1,22 @@
+mod lsp;
+
 #[cfg(test)]
 mod tests {
-    use crate::server::{self, Backend};
-    use crate::utils::test::{
-        assert_server_requests, dir_contains_forc_manifest, doc_comments_dir, e2e_language_dir,
-        e2e_test_dir, get_fixture, runnables_test_dir, sway_workspace_dir, test_fixtures_dir,
+    use crate::{
+        server::Backend,
+        tests::lsp,
+        utils::test::{
+            assert_server_requests, dir_contains_forc_manifest, doc_comments_dir, e2e_language_dir,
+            e2e_test_dir, get_fixture, runnables_test_dir, sway_workspace_dir, test_fixtures_dir,
+        },
     };
     use assert_json_diff::assert_json_eq;
     use serde_json::json;
-    use std::{
-        borrow::Cow,
-        fs,
-        io::Read,
-        path::{Path, PathBuf},
-    };
-    use tower::{Service, ServiceExt};
+    use std::{fs, io::Read, path::PathBuf};
     use tower_lsp::{
-        jsonrpc::{self, Id, Request, Response},
+        jsonrpc::{self, Request, Response},
         lsp_types::*,
-        ExitedError, LspService,
+        LspService,
     };
 
     /// Holds the information needed to check the response of a goto definition request.
@@ -49,176 +48,13 @@ mod tests {
         (uri, sway_program)
     }
 
-    fn build_request_with_id(
-        method: impl Into<Cow<'static, str>>,
-        params: serde_json::Value,
-        id: impl Into<Id>,
-    ) -> Request {
-        Request::build(method).params(params).id(id).finish()
-    }
-
-    async fn call_request(
-        service: &mut LspService<Backend>,
-        req: Request,
-    ) -> Result<Option<Response>, ExitedError> {
-        service.ready().await?.call(req).await
-    }
-
-    async fn initialize_request(service: &mut LspService<Backend>) -> Request {
-        let params = json!({ "capabilities": server::capabilities() });
-        let initialize = build_request_with_id("initialize", params, 1);
-        let response = call_request(service, initialize.clone()).await;
-        let expected =
-            Response::from_ok(1.into(), json!({ "capabilities": server::capabilities() }));
-        assert_json_eq!(expected, response.ok().unwrap());
-        initialize
-    }
-
-    async fn initialized_notification(service: &mut LspService<Backend>) {
-        let initialized = Request::build("initialized").finish();
-        let response = call_request(service, initialized).await;
-        assert_eq!(response, Ok(None));
-    }
-
-    async fn shutdown_request(service: &mut LspService<Backend>) -> Request {
-        let shutdown = Request::build("shutdown").id(1).finish();
-        let response = call_request(service, shutdown.clone()).await;
-        let expected = Response::from_ok(1.into(), json!(null));
-        assert_json_eq!(expected, response.ok().unwrap());
-        shutdown
-    }
-
-    async fn exit_notification(service: &mut LspService<Backend>) {
-        let exit = Request::build("exit").finish();
-        let response = call_request(service, exit.clone()).await;
-        assert_eq!(response, Ok(None));
-    }
-
-    async fn did_open_notification(service: &mut LspService<Backend>, uri: &Url, text: &str) {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-                "languageId": "sway",
-                "version": 1,
-                "text": text,
-            },
-        });
-
-        let did_open = Request::build("textDocument/didOpen")
-            .params(params)
-            .finish();
-        let response = call_request(service, did_open).await;
-        assert_eq!(response, Ok(None));
-    }
-
-    async fn did_change_request(service: &mut LspService<Backend>, uri: &Url) -> Request {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-                "version": 2
-            },
-            "contentChanges": [
-                {
-                    "range": {
-                        "start": {
-                            "line": 1,
-                            "character": 0
-                        },
-                        "end": {
-                            "line": 1,
-                            "character": 0
-                        }
-                    },
-                    "rangeLength": 0,
-                    "text": "\n",
-                }
-            ]
-        });
-        let did_change = Request::build("textDocument/didChange")
-            .params(params)
-            .finish();
-        let response = call_request(service, did_change.clone()).await;
-        assert_eq!(response, Ok(None));
-        did_change
-    }
-
-    async fn did_close_notification(service: &mut LspService<Backend>) {
-        let exit = Request::build("textDocument/didClose").finish();
-        let response = call_request(service, exit.clone()).await;
-        assert_eq!(response, Ok(None));
-    }
-
-    async fn show_ast_request(
-        service: &mut LspService<Backend>,
-        uri: &Url,
-        ast_kind: &str,
-        save_path: Option<Url>,
-    ) -> Request {
-        // The path where the AST will be written to.
-        // If no path is provided, the default path is "/tmp"
-        let save_path = match save_path {
-            Some(path) => path,
-            None => Url::from_file_path(Path::new("/tmp")).unwrap(),
-        };
-        let params = json!({
-            "textDocument": {
-                "uri": uri
-            },
-            "astKind": ast_kind,
-            "savePath": save_path,
-        });
-        let show_ast = build_request_with_id("sway/show_ast", params, 1);
-        let response = call_request(service, show_ast.clone()).await;
-        let expected = Response::from_ok(
-            1.into(),
-            json!({ "uri": format!("{save_path}/{ast_kind}.rs") }),
-        );
-        assert_json_eq!(expected, response.ok().unwrap());
-        show_ast
-    }
-
-    async fn semantic_tokens_request(service: &mut LspService<Backend>, uri: &Url) -> Request {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-            },
-        });
-        let semantic_tokens = build_request_with_id("textDocument/semanticTokens/full", params, 1);
-        let _response = call_request(service, semantic_tokens.clone()).await;
-        semantic_tokens
-    }
-
-    async fn document_symbol_request(service: &mut LspService<Backend>, uri: &Url) -> Request {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-            },
-        });
-        let document_symbol = build_request_with_id("textDocument/documentSymbol", params, 1);
-        let _response = call_request(service, document_symbol.clone()).await;
-        document_symbol
-    }
-
-    fn definition_request(uri: &Url, token_line: i32, token_char: i32, id: i64) -> Request {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-            },
-            "position": {
-                "line": token_line,
-                "character": token_char,
-            }
-        });
-        build_request_with_id("textDocument/definition", params, id)
-    }
-
     async fn definition_check<'a>(
         service: &mut LspService<Backend>,
         go_to: &'a GotoDefintion<'a>,
         id: i64,
     ) -> Request {
-        let definition = definition_request(go_to.req_uri, go_to.req_line, go_to.req_char, id);
-        let response = call_request(service, definition.clone())
+        let definition = lsp::definition_request(go_to.req_uri, go_to.req_line, go_to.req_char, id);
+        let response = lsp::call_request(service, definition.clone())
             .await
             .unwrap()
             .unwrap();
@@ -262,8 +98,11 @@ mod tests {
                 "character": hover_docs.req_char
             }
         });
-        let hover = build_request_with_id("textDocument/hover", params, id);
-        let response = call_request(service, hover.clone()).await.unwrap().unwrap();
+        let hover = lsp::build_request_with_id("textDocument/hover", params, id);
+        let response = lsp::call_request(service, hover.clone())
+            .await
+            .unwrap()
+            .unwrap();
         let value = response.result().unwrap().clone();
         let hover_res: Hover = serde_json::from_value(value).unwrap();
 
@@ -275,207 +114,17 @@ mod tests {
         hover
     }
 
-    async fn format_request(service: &mut LspService<Backend>, uri: &Url) -> Request {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-            },
-            "options": {
-                "tabSize": 4,
-                "insertSpaces": true
-            },
-        });
-        let formatting = build_request_with_id("textDocument/formatting", params, 1);
-        let _response = call_request(service, formatting.clone()).await;
-        formatting
-    }
-
-    async fn highlight_request(service: &mut LspService<Backend>, uri: &Url) -> Request {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-            },
-            "position": {
-                "line": 45,
-                "character": 37
-            }
-        });
-        let highlight = build_request_with_id("textDocument/documentHighlight", params, 1);
-        let response = call_request(service, highlight.clone()).await;
-        let expected = Response::from_ok(
-            1.into(),
-            json!([{
-                    "range": {
-                        "end": {
-                            "character": 41,
-                            "line": 45
-                        },
-                        "start": {
-                            "character": 35,
-                            "line": 45
-                        }
-                    }
-                }
-            ]),
-        );
-        assert_json_eq!(expected, response.ok().unwrap());
-        highlight
-    }
-
-    async fn code_action_request(service: &mut LspService<Backend>, uri: &Url) -> Request {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-            },
-            "range" : {
-                "start":{
-                    "line": 27,
-                    "character": 4
-                },
-                "end":{
-                    "line": 27,
-                    "character": 9
-                },
-            },
-            "context": {
-                "diagnostics": [],
-                "triggerKind": 2
-            }
-        });
-        let code_action = build_request_with_id("textDocument/codeAction", params, 1);
-        let response = call_request(service, code_action.clone()).await;
-        let uri_string = uri.to_string();
-        let expected = Response::from_ok(
-            1.into(),
-            json!([{
-                "data": uri,
-                "edit": {
-                  "changes": {
-                    uri_string: [
-                      {
-                        "newText": "\nimpl FooABI for Contract {\n    /// This is the `main` method on the `FooABI` abi\n    fn main() -> u64 {}\n}\n",
-                        "range": {
-                          "end": {
-                            "character": 0,
-                            "line": 31
-                          },
-                          "start": {
-                            "character": 0,
-                            "line": 31
-                          }
-                        }
-                      }
-                    ]
-                  }
-                },
-                "kind": "refactor",
-                "title": "Generate impl for contract"
-            }]),
-        );
-        assert_json_eq!(expected, response.ok().unwrap());
-        code_action
-    }
-
-    async fn code_lens_request(service: &mut LspService<Backend>, uri: &Url) -> Request {
-        let params = json!({
-            "textDocument": {
-                "uri": uri,
-            },
-        });
-        let code_lens = build_request_with_id("textDocument/codeLens", params, 1);
-        let response = call_request(service, code_lens.clone()).await;
-        let actual_results = response
-            .unwrap()
-            .unwrap()
-            .into_parts()
-            .1
-            .ok()
-            .unwrap()
-            .as_array()
-            .unwrap()
-            .clone();
-        let expected_results = vec![
-            json!({
-              "command": {
-                "arguments": [
-                  {
-                    "name": "test_bar"
-                  }
-                ],
-                "command": "sway.runTests",
-                "title": "▶︎ Run Test"
-              },
-              "range": {
-                "end": {
-                  "character": 7,
-                  "line": 11
-                },
-                "start": {
-                  "character": 0,
-                  "line": 11
-                }
-              }
-            }),
-            json!({
-              "command": {
-                "arguments": [
-                  {
-                    "name": "test_foo"
-                  }
-                ],
-                "command": "sway.runTests",
-                "title": "▶︎ Run Test"
-              },
-              "range": {
-                "end": {
-                  "character": 7,
-                  "line": 6
-                },
-                "start": {
-                  "character": 0,
-                  "line": 6
-                }
-              }
-            }),
-            json!({
-              "command": {
-                "command": "sway.runScript",
-                "title": "▶︎ Run"
-              },
-              "range": {
-                "end": {
-                  "character": 7,
-                  "line": 2
-                },
-                "start": {
-                  "character": 3,
-                  "line": 2
-                }
-              }
-            }),
-        ];
-
-        assert_eq!(actual_results.len(), expected_results.len());
-        for expected in expected_results.iter() {
-            assert!(
-                actual_results.contains(expected),
-                "Expected {actual_results:?} to contain {expected:?}"
-            );
-        }
-        code_lens
-    }
-
     async fn init_and_open(service: &mut LspService<Backend>, entry_point: PathBuf) -> Url {
-        let _ = initialize_request(service).await;
-        initialized_notification(service).await;
+        let _ = lsp::initialize_request(service).await;
+        lsp::initialized_notification(service).await;
         let (uri, sway_program) = load_sway_example(entry_point);
-        did_open_notification(service, &uri, &sway_program).await;
+        lsp::did_open_notification(service, &uri, &sway_program).await;
         uri
     }
 
     async fn shutdown_and_exit(service: &mut LspService<Backend>) {
-        let _ = shutdown_request(service).await;
-        exit_notification(service).await;
+        let _ = lsp::shutdown_request(service).await;
+        lsp::exit_notification(service).await;
     }
 
     // This method iterates over all of the examples in the e2e langauge should_pass dir
@@ -488,8 +137,8 @@ mod tests {
         let (mut service, _) = LspService::build(Backend::new)
             .custom_method("sway/show_ast", Backend::show_ast)
             .finish();
-        let _ = initialize_request(&mut service).await;
-        initialized_notification(&mut service).await;
+        let _ = lsp::initialize_request(&mut service).await;
+        lsp::initialized_notification(&mut service).await;
 
         let ast_folder = dirs::home_dir()
             .expect("could not get users home directory")
@@ -521,10 +170,12 @@ mod tests {
 
                 let example_dir = Some(Url::from_file_path(example_dir).unwrap());
                 let (uri, sway_program) = load_sway_example(manifest_dir.join("src/main.sw"));
-                did_open_notification(&mut service, &uri, &sway_program).await;
-                let _ = show_ast_request(&mut service, &uri, "lexed", example_dir.clone()).await;
-                let _ = show_ast_request(&mut service, &uri, "parsed", example_dir.clone()).await;
-                let _ = show_ast_request(&mut service, &uri, "typed", example_dir).await;
+                lsp::did_open_notification(&mut service, &uri, &sway_program).await;
+                let _ =
+                    lsp::show_ast_request(&mut service, &uri, "lexed", example_dir.clone()).await;
+                let _ =
+                    lsp::show_ast_request(&mut service, &uri, "parsed", example_dir.clone()).await;
+                let _ = lsp::show_ast_request(&mut service, &uri, "typed", example_dir).await;
             }
         }
         shutdown_and_exit(&mut service).await;
@@ -533,22 +184,22 @@ mod tests {
     #[tokio::test]
     async fn initialize() {
         let (mut service, _) = LspService::new(Backend::new);
-        let _ = initialize_request(&mut service).await;
+        let _ = lsp::initialize_request(&mut service).await;
     }
 
     #[tokio::test]
     async fn initialized() {
         let (mut service, _) = LspService::new(Backend::new);
-        let _ = initialize_request(&mut service).await;
-        initialized_notification(&mut service).await;
+        let _ = lsp::initialize_request(&mut service).await;
+        lsp::initialized_notification(&mut service).await;
     }
 
     #[tokio::test]
     async fn initializes_only_once() {
         let (mut service, _) = LspService::new(Backend::new);
-        let initialize = initialize_request(&mut service).await;
-        initialized_notification(&mut service).await;
-        let response = call_request(&mut service, initialize).await;
+        let initialize = lsp::initialize_request(&mut service).await;
+        lsp::initialized_notification(&mut service).await;
+        let response = lsp::call_request(&mut service, initialize).await;
         let err = Response::from_error(1.into(), jsonrpc::Error::invalid_request());
         assert_eq!(response, Ok(Some(err)));
     }
@@ -556,21 +207,21 @@ mod tests {
     #[tokio::test]
     async fn shutdown() {
         let (mut service, _) = LspService::new(Backend::new);
-        let _ = initialize_request(&mut service).await;
-        initialized_notification(&mut service).await;
-        let shutdown = shutdown_request(&mut service).await;
-        let response = call_request(&mut service, shutdown).await;
+        let _ = lsp::initialize_request(&mut service).await;
+        lsp::initialized_notification(&mut service).await;
+        let shutdown = lsp::shutdown_request(&mut service).await;
+        let response = lsp::call_request(&mut service, shutdown).await;
         let err = Response::from_error(1.into(), jsonrpc::Error::invalid_request());
         assert_eq!(response, Ok(Some(err)));
-        exit_notification(&mut service).await;
+        lsp::exit_notification(&mut service).await;
     }
 
     #[tokio::test]
     async fn refuses_requests_after_shutdown() {
         let (mut service, _) = LspService::new(Backend::new);
-        let _ = initialize_request(&mut service).await;
-        let shutdown = shutdown_request(&mut service).await;
-        let response = call_request(&mut service, shutdown).await;
+        let _ = lsp::initialize_request(&mut service).await;
+        let shutdown = lsp::shutdown_request(&mut service).await;
+        let response = lsp::call_request(&mut service, shutdown).await;
         let err = Response::from_error(1.into(), jsonrpc::Error::invalid_request());
         assert_eq!(response, Ok(Some(err)));
     }
@@ -586,7 +237,7 @@ mod tests {
     async fn did_close() {
         let (mut service, _) = LspService::new(Backend::new);
         let _ = init_and_open(&mut service, e2e_test_dir().join("src/main.sw")).await;
-        did_close_notification(&mut service).await;
+        lsp::did_close_notification(&mut service).await;
         shutdown_and_exit(&mut service).await;
     }
 
@@ -594,7 +245,7 @@ mod tests {
     async fn did_change() {
         let (mut service, _) = LspService::new(Backend::new);
         let uri = init_and_open(&mut service, doc_comments_dir().join("src/main.sw")).await;
-        let _ = did_change_request(&mut service, &uri).await;
+        let _ = lsp::did_change_request(&mut service, &uri).await;
         shutdown_and_exit(&mut service).await;
     }
 
@@ -612,7 +263,7 @@ mod tests {
             def_path: uri.as_str(),
         };
         let _ = definition_check(&mut service, &go_to, 1).await;
-        let _ = did_change_request(&mut service, &uri).await;
+        let _ = lsp::did_change_request(&mut service, &uri).await;
         go_to.def_line = 20;
         definition_check_with_req_offset(&mut service, &mut go_to, 45, 24, 2).await;
         shutdown_and_exit(&mut service).await;
@@ -625,7 +276,7 @@ mod tests {
             .finish();
 
         let uri = init_and_open(&mut service, e2e_test_dir().join("src/main.sw")).await;
-        let _ = show_ast_request(&mut service, &uri, "typed", None).await;
+        let _ = lsp::show_ast_request(&mut service, &uri, "typed", None).await;
         shutdown_and_exit(&mut service).await;
     }
 
@@ -1288,32 +939,32 @@ mod tests {
 
     lsp_capability_test!(
         semantic_tokens,
-        semantic_tokens_request,
+        lsp::semantic_tokens_request,
         doc_comments_dir().join("src/main.sw")
     );
     lsp_capability_test!(
         document_symbol,
-        document_symbol_request,
+        lsp::document_symbol_request,
         doc_comments_dir().join("src/main.sw")
     );
     lsp_capability_test!(
         format,
-        format_request,
+        lsp::format_request,
         doc_comments_dir().join("src/main.sw")
     );
     lsp_capability_test!(
         highlight,
-        highlight_request,
+        lsp::highlight_request,
         doc_comments_dir().join("src/main.sw")
     );
     lsp_capability_test!(
         code_action,
-        code_action_request,
+        lsp::code_action_request,
         doc_comments_dir().join("src/main.sw")
     );
     lsp_capability_test!(
         code_lens,
-        code_lens_request,
+        lsp::code_lens_request,
         runnables_test_dir().join("src/main.sw")
     );
 }
